# Conflicts:
#	sonic-shelf-admin/src/api/data-format-guide.js
#	sonic-shelf-admin/src/api/migration-guide.js

歌手管理页面，用户管理页面完成

+++ sonic-shelf-server/src/main/java/com/zhongxin/sonicshelf/entity/User.java
@@ -30,7 +30,7 @@ public class User implements UserDetails {
 
     // ========== 状态字段 ==========
     private Boolean emailVerified = false;
-    private Integer status = 1;
+    private byte status = 1;
 
     // ========== 统计字段 ==========
     private Integer followerCount = 0;
